+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Note: This file was started on 2023-04-22. Any changes before then are not captu
 - Removed the femboy easter egg. Was a joke that overstayed its welcome
 - Removed some pointless conditions from the the community roles functionality
 - Tidied up the !enlist function
+- Updated frontend messaging to replace 'users' with 'members'
 
 ## 2023-12-01
 - Introduced error handling into dm_message listener and improved logging